refactor: Today の Server Actions を責務別に分割
- src/app/today/actions.ts を削除し actions/*.ts に実装を移す
- lib/constants/meal.ts と lib/date.ts を追加し参照を集約
- ドキュメント内の旧パスを更新

closes #215

Made-with: Cursor
