+++ src/app/today/page.tsx
@@ -2,7 +2,8 @@ import { getSupabaseAuthForRequest } from "@/lib/supabase/request-auth";
 import { getMealTypeForTimeZone } from "@/lib/meal-timezone";
 import { redirect } from "next/navigation";
 import TodayClient from "./TodayClient";
-import { getOrCreateSnapshotRestaurant, fetchFavoriteGroupsPayload } from "./actions";
+import { getOrCreateSnapshotRestaurant } from "./actions/restaurant";
+import { fetchFavoriteGroupsPayload } from "./actions/favorites";
 import type { FoodLogEntry, MenuItem, Restaurant, UserSettings, TodayConsumed } from "@/types/database";
 import { normalizeUserSettings } from "@/lib/diet-phase";
 import { loadPresets } from "@/lib/presets-server";
